move scheduler classes out of models
